Moved StackDepthExceptionHelper out into separate file, and make a scala.js version that always return 0.
