+++ scalatest.js/src/main/scala/org/scalatest/exceptions/StackDepthExceptionHelper.scala
@@ -1,0 +1,11 @@
+package org.scalatest.exceptions
+
+private[scalatest] object StackDepthExceptionHelper {
+
+  def getStackDepth(stackTrace: Array[StackTraceElement], fileName: String, methodName: String, adjustment: Int = 0): Int =
+    0
+
+  def getStackDepthFun(fileName: String, methodName: String, adjustment: Int = 0): (StackDepthException => Int) = { sde =>
+    getStackDepth(sde.getStackTrace, fileName, methodName, adjustment)
+  }
+}
