+++ scalatest.js/src/test/scala/test/ExampleSuite.scala
@@ -5,7 +5,9 @@ import org.scalatest._
 class ExampleSuite extends FunSuite {
 
   test("test 1") {
-    assert(1 == 1)
+    //System.getProperty("line.separator")
+    //val a = 1
+    assert(1 == 2)
   }
 
 }
